condaenv.25mi2wga.requirements.txt
transformers>=4.39,<4.50
accelerate
bitsandbytes
deepspeed
TTS==0.22.0
chromadb==1.0.17
ctransformers==0.2.27
faiss-cpu==1.8.0
langchain==0.3.27
langchain-openai==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-huggingface==0.3.1
pikepdf==8.15.0
pillow_heif==0.16.0
PyMuPDF==1.23.26
python-box==7.1.1
rank-bm25==0.2.2
sentence-transformers==2.6.0
simsimd==3.3.0
streamlit==1.33.0
umap-learn==0.5.5
unstructured==0.13.2
unstructured-inference==0.7.27
pdf2image==1.17.0
pdfminer.six==20231228
opencv-python==4.9.0.80
soundfile==0.12.1
mutagen==1.47.0
pygame-ce==2.4.1
pygame==2.5.2
SpeechRecognition==3.10.3
pyttsx3==2.90
pyaudio==0.2.14
pyarrow==11.0.0
sounddevice==0.4.6
faster-whisper == 1.2.0
librosa
torchinfo
multiprocess
auto-gptq
tqdm>=4.65.0
PyMuPDF==1.23.26